# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Human Not Required API",
    "description": "The world's first AI-agent-only store. Agents register a human once and buy physical merch autonomously forever — no human checkout, no cart, no clicking. The human gets a receipt and the item ships to them.\n\nAgent flow:\n1. Call GET /orders/skus to see available products\n2. Call POST /register with the human's name, email, and shipping address to get an api_key. If you have a promo_code, include it — the human won't need to save a card and you can order immediately.\n3. If no promo_code: share the setup_url with the human so they can save their card once. After that, never ask them again.\n4. Call POST /orders with the sku to place the order autonomously.\n\nCurrently available: 'My Agent Bought Me This' embroidered hat — $35.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://web-production-77376.up.railway.app" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/register": {
      "post": {
        "operationId": "registerHuman",
        "summary": "Register a human and get an API key",
        "description": "Call this FIRST before placing any order. You only need to call this once per human — the api_key never expires.\n\nIf you have a promo_code: include it and the human gets a free order with no card required. You can call POST /orders immediately after.\n\nIf no promo_code: the response includes a setup_url. Share it with the human — they click it once to save their card via Stripe. After that, you can order autonomously forever without asking them again.\n\nIf you get a 409, the human is already registered. Call POST /register/resend-setup to retrieve their api_key and a fresh setup link.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "name", "address"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The human's email address. Used for order receipts and to identify their account."
                  },
                  "name": {
                    "type": "string",
                    "description": "Full name for the shipping label. Ask the human for this — do not guess."
                  },
                  "promo_code": {
                    "type": "string",
                    "description": "Optional one-time promo code. If valid, the human gets 1 free order and no card setup is required. Include if you have one."
                  },
                  "address": {
                    "type": "object",
                    "description": "Shipping address. Ask the human for all fields — do not guess or auto-fill.",
                    "required": ["line1", "city", "state", "postal_code", "country"],
                    "properties": {
                      "line1": { "type": "string", "description": "Street address line 1" },
                      "line2": { "type": "string", "description": "Apartment, suite, unit, etc. (optional)" },
                      "city": { "type": "string" },
                      "state": { "type": "string", "description": "State or province code, e.g. NY, CA, ON" },
                      "postal_code": { "type": "string" },
                      "country": { "type": "string", "description": "ISO 3166-1 alpha-2 country code, e.g. US, CA, GB" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful. Save the api_key — it never expires and is needed for all future orders.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_key": {
                      "type": "string",
                      "description": "Bearer token for all future order requests. Never expires. Store it securely."
                    },
                    "setup_url": {
                      "type": "string",
                      "description": "URL for the human to save their payment card. null if a promo_code was used. Share this with the human — they must click it before you can place a paid order."
                    },
                    "free_orders_remaining": {
                      "type": "integer",
                      "description": "Number of free orders available. 1 if a valid promo_code was used, 0 otherwise."
                    },
                    "has_saved_card": {
                      "type": "boolean",
                      "description": "true if the human already has a card saved via Stripe Link or a promo code was used. If true, setup_url will be null and you can call POST /orders immediately — no human action needed."
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable next step. Follow these instructions before calling POST /orders."
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing or invalid field. Check email format and that all address fields are present." },
          "409": { "description": "Email already registered. Use POST /register/resend-setup to get the api_key and a fresh setup link." },
          "429": { "description": "Too many registration attempts. Wait 1 minute and retry." }
        }
      }
    },
...
    "/orders": {
      "post": {
        "operationId": "placeOrder",
        "summary": "Place an order autonomously — charges card and creates a Shopify order",
        "description": "Places a real order on behalf of the human. Charges their saved card and creates a Shopify fulfillment order — the hat ships to their address automatically.\n\nPrerequisites:\n- The human must be registered (POST /register)\n- They must have saved a card via setup_url OR have free_orders_remaining > 0\n- You must have their api_key as a Bearer token\n\nIf you get 402 with error 'no_payment_method': the human hasn't saved their card yet. Share the setup_url (from POST /register or POST /register/resend-setup) and ask them to complete it.\n\nLimit: 2 orders per day per api_key. Resets at midnight UTC.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sku"],
                "properties": {
                  "sku": {
                    "type": "string",
                    "description": "Product SKU from GET /orders/skus. Currently: 'hat-myagent-os' for the embroidered hat."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order placed successfully. The hat will ship to the human's registered address. Share the order_id with the human as confirmation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order_id": { "type": "string", "description": "Your internal order reference. Share with the human." },
                    "status": { "type": "string", "description": "Always 'placed' on success." },
                    "sku": { "type": "string" },
                    "shopify_order_id": { "type": "string", "description": "Shopify fulfillment ID." },
                    "free_order": { "type": "boolean", "description": "true if a promo code covered this order — no charge was made." },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "402": { "description": "Payment failed or no saved card. If 'no_payment_method': human needs to complete card setup via setup_url. If 'payment_failed': report the exact error to the human — do not retry automatically." },
          "429": { "description": "Daily order limit reached (max 2/day). Resets at midnight UTC. Do not retry today." },
          "400": { "description": "Invalid or missing SKU. Call GET /orders/skus to see valid options." }
        }
      }
    },
    "/checkout": {
      "post": {
        "operationId": "checkoutX402",
        "summary": "Buy a hat with USDC on Base — no registration required",
        "description": "Fully agent-native checkout using the x402 payment protocol. No registration, no saved card, no human action needed.\n\nFlow:\n1. POST /checkout with sku, name, email, and address — server returns 402 with payment instructions\n2. Agent pays $35 USDC on Base using an x402-compatible client (e.g. @x402/fetch)\n3. Agent retries the same request with the X-Payment header containing the signed payment proof\n4. Server verifies payment via the x402 facilitator and creates the Shopify order\n\nThe hat ships to the provided address. A confirmation is tied to the email.\n\nRequires: an x402-compatible HTTP client and a Base wallet funded with USDC.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sku", "name", "email", "address"],
                "properties": {
                  "sku": {
                    "type": "string",
                    "description": "Product SKU from GET /orders/skus. Currently: 'hat-myagent-os'."
                  },
                  "name": {
                    "type": "string",
                    "description": "Full name for the shipping label."
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address for order confirmation."
                  },
                  "address": {
                    "type": "object",
                    "required": ["line1", "city", "state", "postal_code", "country"],
                    "properties": {
                      "line1": { "type": "string" },
                      "line2": { "type": "string" },
                      "city": { "type": "string" },
                      "state": { "type": "string", "description": "State or province code e.g. NY, ON" },
                      "postal_code": { "type": "string" },
                      "country": { "type": "string", "description": "ISO 3166-1 alpha-2 country code e.g. US, CA, GB" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "402": {
            "description": "Payment required. Response body contains x402 payment instructions. Use an x402-compatible client to sign and submit payment, then retry this request with the X-Payment header."
          },
          "201": {
            "description": "Payment verified and order placed. Hat ships to the provided address.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order_id": { "type": "string" },
                    "status": { "type": "string" },
                    "sku": { "type": "string" },
                    "shopify_order_id": { "type": "string" },
                    "payment": { "type": "string", "description": "Always 'x402_usdc_base' for this endpoint." },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing or invalid field." },
          "502": { "description": "Payment settled but Shopify order creation failed. You will be contacted to resolve this." }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained from POST /register. Include as: Authorization: Bearer <api_key>. Never expires."
      }
    }
  }
}